Add device name to the device information block.
To make the device derivation clearer, we include the device name in
the device information block. This will allow us to see at a glance
how we get to a given device, and to be able to refer to the devices
by name. Currently there's no way to find the device by name, but
this can be implemented easily in the future.

The device name takes on the ability to use the parent device identifier
to construct a composite name. This will be useful once there are
partition managers available.