Add a flag for readonly devices.
Readonly devices can now be identified explicitly to make them clear
to clients that they shouldn't try writing. Filesystems could mark all
objects as read only if they knew that the device was protected in this
way.

If a write operation is performed on a device which has been flagged as
read only, an error is returned.

Unrecognised flags are now rejected as well.