Add miscop (#8)
* Add stub definitions for MiscOp calls.

We want to be able to do PowerEject. We're going to need ReadCapacity.
There's going to need to be operations like SMART (or at least health)
information reading and probably TRIM and others. MiscOp allows some
of these operations to be fed through.

* Update the stub MiscOp with an actual call to the MiscOp entrypoint.

The MiscOp entrypoint is now called when you call the SWI.

* Move some of the MiscOp definitions around.

Solidify some of the MiscOp definitions in documentation.